requirements.txt
python-dotenv
requests
msal
traceback
mimetypes